requirements.txt
boto==2.48.0
bumpversion==0.5.3
coverage==4.4.1
Django==1.10.7
django-filter==1.0.4
django-mailbox==4.5.4
django-storages==1.6.4
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
pyexcel==0.5.3
pyexcel-xlsx==0.4.1
PyJWT
PyPDF2==1.26.0
python-dateutil==2.6.1
pytz==2017.2
raven==6.1.0
requests==2.18.2
router==0.1
drf-writable-nested==0.2.0
urllib3==1.22
django-revproxy==0.9.12
selenium==3.3.0
xvfbwrapper==0.2.8
factory-boy==2.8.1
Faker==0.7.10
beautifulsoup4==4.4.1
Pillow==2.9.0
django-taggit==0.22.1
django-taggit-serializer==0.1.5
django-dynamic-preferences==1.5
jinja2==2.10
django-rest-auth==0.9.3
django-post-office==3.0.4
django-logentry-admin==1.0.4
django-admin-honeypot==1.1.0
pyxero
psycopg2
pycrypto==2.6.1
